+++ ser/train.py
@@ -1,0 +1,51 @@
+
+
+import torch
+import torch.nn.functional as F
+from ser.data import get_data_loaders
+from ser.model import Net
+from torch import optim
+
+
+def training_function(epochs, batch_size, learning_rate):
+
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+
+    # save the parameters!
+    # load model
+    model = Net().to(device)
+
+    # setup params
+    optimizer = optim.Adam(model.parameters(), lr=learning_rate)
+
+    training_dataloader, validation_dataloader = get_data_loaders(batch_size)
+    for epoch in range(epochs):
+        for i, (images, labels) in enumerate(training_dataloader):
+            images, labels = images.to(device), labels.to(device)
+            model.train()
+            optimizer.zero_grad()
+            output = model(images)
+            loss = F.nll_loss(output, labels)
+            loss.backward()
+            optimizer.step()
+            print(
+                f"Train Epoch: {epoch} | Batch: {i}/{len(training_dataloader)} "
+                f"| Loss: {loss.item():.4f}"
+            )
+            # validate
+            val_loss = 0
+            correct = 0
+            with torch.no_grad():
+                for images, labels in validation_dataloader:
+                    images, labels = images.to(device), labels.to(device)
+                    model.eval()
+                    output = model(images)
+                    val_loss += F.nll_loss(output, labels, reduction="sum").item()
+                    pred = output.argmax(dim=1, keepdim=True)
+                    correct += pred.eq(labels.view_as(pred)).sum().item()
+                val_loss /= len(validation_dataloader.dataset(batch_size))
+                val_acc = correct / len(validation_dataloader.dataset(batch_size))
+
+                print(
+                    f"Val Epoch: {epoch} | Avg Loss: {val_loss:.4f} | Accuracy: {val_acc}"
+                )
