+++ ser/transforms.py
@@ -1,0 +1,10 @@
+from torchvision import transforms
+
+def get_transforms():
+    # torch transforms
+    ts = transforms.Compose(
+        [transforms.ToTensor(), transforms.Normalize((0.5,), (0.5,))]
+    )
+    return ts
+
+
